Extract IHV Interactive Cards tab and DRY provider card props
Move the cards-tab pass UI into HardwarePassCards.tsx (mirroring the matrix extract). Share one providerCardProps object across local/export HardwareProviderCard sections to clear the CodeFactor duplicate (#159).
